requirements.txt
# Patent Guard v2.0 - Requirements
# Optimized for Octen-Embedding-8B with Intel iGPU (IPEX-LLM INT4)
# Environment: patent-guard (Conda)

# =============================================================================
# Core ML Framework (Intel Optimized)
# =============================================================================
# NOTE: PyTorch for Intel XPU should be installed via conda first:
#   conda install pytorch torchvision torchaudio cpuonly -c pytorch
# Then install Intel Extension:
#   pip install --pre --upgrade ipex-llm[xpu] --extra-index-url https://pytorch-extension.intel.com/release-whl/stable/xpu/us/

transformers>=4.38.0
accelerate>=0.27.0

# =============================================================================
# Intel XPU Optimization (replaces bitsandbytes)
# =============================================================================
# Install separately with Intel index:
# pip install --pre --upgrade ipex-llm[xpu] --extra-index-url https://pytorch-extension.intel.com/release-whl/stable/xpu/us/

# =============================================================================
# Tokenizer Dependencies
# =============================================================================
sentencepiece>=0.2.0
protobuf>=4.25.0

# =============================================================================
# Vector Operations
# =============================================================================
scipy>=1.12.0
numpy>=1.26.0

# =============================================================================
# Utilities
# =============================================================================
tqdm>=4.66.0

# =============================================================================
# Async Support
# =============================================================================
aiofiles>=23.2.1

# =============================================================================
# Vector Database (Milvus)
# =============================================================================
pymilvus>=2.4.0

# =============================================================================
# LangGraph & RAG Components
# =============================================================================
langgraph>=0.1.0
langchain>=0.2.0
langchain-google-genai>=1.0.0

# =============================================================================
# BigQuery (Data Source)
# =============================================================================
google-cloud-bigquery>=3.20.0

# =============================================================================
# Environment & Secrets Management
# =============================================================================
python-dotenv>=1.0.0

# =============================================================================
# OpenAI API (for Self-RAG analysis)
# =============================================================================
openai>=1.12.0
